+++ darts/utils/ts_utils.py
@@ -3,9 +3,7 @@ Additional util functions
 -------------------------
 """
 
-from typing import Any, List, Optional, Sequence, Union
-
-import numpy as np
+from typing import List, Optional, Sequence, Union
 
 from darts import TimeSeries
 from darts.logging import get_logger, raise_log
@@ -52,11 +50,12 @@ def retain_period_common_to_all(series: List[TimeSeries]) -> List[TimeSeries]:
 
 
 def series2seq(
-    ts: Optional[Union[Any, Sequence[Any], Sequence[Sequence[Any]]]],
+    ts: Optional[
+        Union[TimeSeries, Sequence[TimeSeries], Sequence[Sequence[TimeSeries]]]
+    ],
     seq_type_out: int = 1,
-    is_numeric: bool = False,
     nested: bool = False,
-) -> Optional[Union[Any, Sequence[Any], Sequence[Sequence[Any]]]]:
+) -> Optional[Union[TimeSeries, Sequence[TimeSeries], Sequence[Sequence[TimeSeries]]]]:
     """If possible, converts `ts` into the desired sequence type `seq_type_out`. Otherwise, returns the
     original `ts`.
 
@@ -70,9 +69,9 @@ def series2seq(
         - 0: `TimeSeries` (e.g. a single series)
         - 1: sequence of `TimeSeries` (e.g. multiple series)
         - 2: sequence of sequences of `TimeSeries` (e.g. historical forecasts output)
-    is_numeric
-        Whether to look for a numeric value as the innermost value. If `False`, looks
-        for a `TimeSeries` object.
+    nested
+        Only applies with `seq_type_out=2` and `ts` having a sequence type `1`. In this case, wrap each element in
+        `ts` in a list ([ts1, ts2] -> [[ts1], [ts2]]).
 
     Raises
     ------
@@ -90,7 +89,7 @@ def series2seq(
             logger=logger,
         )
 
-    seq_type_in = get_series_seq_type(ts, is_numeric=is_numeric)
+    seq_type_in = get_series_seq_type(ts)
 
     if seq_type_out == seq_type_in:
         return ts
@@ -114,20 +113,11 @@ def series2seq(
             # [ts1, ts2] -> [[ts1], [ts2]]
             return [[ts_] for ts_ in ts]
     elif seq_type_in == 2 and seq_type_out == 0 and n_series == 1:
-        if not nested:
-            # [[ts]] -> [ts]
-            return ts[0]
-        else:
-            # either [[ts]] -> ts or [[ts1, ts2]] -> [ts1, ts2]
-            return series2seq(ts[0], seq_type_out=seq_type_out, is_numeric=is_numeric)
+        # [[ts]] -> [ts]
+        return ts[0]
     elif seq_type_in == 2 and seq_type_out == 1 and n_series == 1:
-        if not nested:
-            # [[ts1, ts2]] -> [[ts1, ts2]]
-            return ts
-        else:
-            # seq_type_in `1` and `2` are both representing time series axis
-            # [[ts1, ts2]] -> [ts1, ts2]
-            return ts[0]
+        # [[ts1, ts2]] -> [[ts1, ts2]]
+        return ts
     else:
         # ts -> ts
         return ts
@@ -176,7 +166,6 @@ def get_single_series(
 
 def get_series_seq_type(
     ts: Union[TimeSeries, Sequence[TimeSeries], Sequence[Sequence[TimeSeries]]],
-    is_numeric: bool = False,
 ) -> int:
     """Returns the sequence type of `ts`.
 
@@ -188,34 +177,23 @@ def get_series_seq_type(
     ----------
     ts
         The input series to get the sequence type from.
-    is_numeric
-        Whether to look for a numeric value as the innermost value. If `False`, looks
-        for a `TimeSeries` object.
 
     Raises
     ------
     ValueError
         If `ts` does not have one of the expected sequence types.
     """
-
-    def is_inner_value(val) -> bool:
-        if not is_numeric:
-            return isinstance(val, TimeSeries)
-        else:
-            return np.issubdtype(type(val), np.number)
-
-    if is_inner_value(ts):
+    if isinstance(ts, TimeSeries):
         return 0
-    elif is_inner_value(ts[0]):
+    elif isinstance(ts[0], TimeSeries):
         return 1
-    elif is_inner_value(ts[0][0]):
+    elif isinstance(ts[0][0], TimeSeries):
         return 2
     else:
-        ts_type = "TimeSeries" if not is_numeric else "np.number"
         raise_log(
             ValueError(
-                f"`ts` must be of type `{ts_type}`, `Sequence[{ts_type}]`, or "
-                f"`Sequence[Sequence[{ts_type}]]`"
+                "input series must be of type `TimeSeries`, `Sequence[TimeSeries]`, or "
+                "`Sequence[Sequence[TimeSeries]]`"
             ),
             logger=logger,
         )